Separating contracts.
