+++ src/Contracts/Blockable.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rennokki\Befriended\Contracts;
+
+interface Blockable
+{
+    public function blockers($model = null);
+}
