+++ src/Contracts/Blocker.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Rennokki\Befriended\Contracts;
+
+interface Blocker
+{
+    public function blocking($model = null);
+
+    public function isBlocking($model);
+
+    public function block($model);
+
+    public function unblock($model);
+}
