+++ src/Contracts/Blocking.php
@@ -2,15 +2,10 @@
 
 namespace Rennokki\Befriended\Contracts;
 
-interface Blocking
+use Rennokki\Befriended\Contracts\Blockable;
+use Rennokki\Befriended\Contracts\Blocker;
+
+interface Blocking extends Blockable, Blocker
 {
-    public function blockers($model = null);
-
-    public function blocking($model = null);
-
-    public function isBlocking($model);
-
-    public function block($model);
-
-    public function unblock($model);
+    //
 }
