+++ src/Contracts/Followable.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rennokki\Befriended\Contracts;
+
+interface Followable
+{
+    public function followers($model = null);
+}
